:recycle: admin: rename search bar component
