--- admin/app/components/complementary-certifications/target-profiles/attach-target-profile.hbs
@@ -1,35 +1,0 @@
-<div class="card">
-  <div class="card__title">1. Renseigner le nouveau profil cible à rattacher</div>
-  <div class="card__content">
-    <form class="complementary-certification-attach-target-profile__form">
-      <div class="complementary-certification-attach-target-profile-form-search">
-        <PixSearchInput
-          @id="search-input"
-          @label="ID du profil cible"
-          @ariaLabel="Rechercher un profil cible à rattacher à la certification complémentaire"
-          @placeholder="Exemple: 3"
-          @debounceTimeInMs={{this.debounce}}
-          @triggerFiltering={{this.onSearchValueInput}}
-          autocomplete="off"
-        />
-        {{#if @options.length}}
-          <ul role="listbox" class="complementary-certification-attach-target-profile-form-search__results">
-            {{#each @options as |option|}}
-              <li
-                class="complementary-certification-attach-target-profile-form-search__results__option"
-                role="option"
-                aria-selected="false"
-                tabindex="0"
-                {{on "click" (fn this.onSelectTargetProfile option)}}
-                {{on-enter-action (fn this.onSelectTargetProfile option)}}
-                {{on-space-action (fn this.onSelectTargetProfile option)}}
-              >
-                {{option.label}}
-              </li>
-            {{/each}}
-          </ul>
-        {{/if}}
-      </div>
-    </form>
-  </div>
-</div>
