+++ admin/app/styles/app.scss
@@ -45,6 +45,7 @@
 @import 'components/certification-center-invitations';
 @import 'components/complementary-certifications/details';
 @import 'components/complementary-certifications/attach-target-profile';
+@import 'components/complementary-certifications/search-bar';
 @import 'components/profile';
 @import 'components/certified-profile';
 @import 'components/confirm-popup';
