+++ admin/app/styles/components/complementary-certifications/search-bar.scss
@@ -1,0 +1,54 @@
+
+.attach-target-profile-search {
+    display: flex;
+    gap: 10px;
+    align-items: flex-end;
+    padding: $pix-spacing-s 0;
+
+  &__input {
+    min-width: 40%;
+  }
+
+  &__results {
+    position: absolute;
+    z-index: 200;
+    width: inherit;
+    max-height: 12.5rem;
+    padding: $pix-spacing-xs $pix-spacing-s;
+    overflow-y: auto;
+    white-space: nowrap;
+    list-style-type: none;
+    background-color: $pix-neutral-0;
+    border-top: none;
+    border-radius: 0 0 $pix-spacing-xxs $pix-spacing-xxs;
+    box-shadow: 0 6px 12px rgb(7 20 46 / 8%);
+    transition: all 0.1s ease-in-out;
+
+    &__option:hover,
+    &__option:focus {
+      background-color: $pix-neutral-10;
+      outline: none;
+      cursor: pointer;
+    }
+
+    &::-webkit-scrollbar {
+      width: 0.5rem;
+    }
+
+    &::-webkit-scrollbar-track {
+      background: $pix-neutral-20;
+      border: 1px solid $pix-neutral-20;
+      border-radius: $pix-spacing-xxs;
+    }
+
+    &::-webkit-scrollbar-thumb {
+      width: 0.375rem;
+      background: $pix-neutral-30;
+      border-radius: $pix-spacing-xxs;
+    }
+
+    &::-webkit-scrollbar-thumb:hover {
+      background: $pix-neutral-35;
+    }
+  }
+}
